Search MGnify database for studies, samples and runs

Search MGnify database for studies, samples and runs

doQuery(x, ...)

# S4 method for MgnifyClient
doQuery(
  x,
  type = c("studies", "samples", "runs", "analyses"),
  accession = NULL,
  as.df = TRUE,
  max.hits = 200,
  use.cache = FALSE,
  ...
)

Arguments

x

A MgnifyClient object.

...

Remaining parameter key/value pairs may be supplied to filter the returned values. Available options differ between types. See discussion above for details.

type

A single character value specifying the type of objects to query. Must be one of the following options: studies, samples, runs or analyses. (By default: type = "studies")

accession

A single character value or a vector of character values specifying MGnify accession identifiers (of type type) or NULL. When NULL, all results defined by other parameters are retrieved. (By default: accession = NULL)

as.df

A single boolean value specifying whether to return the results as a data.frame or leave as a nested list. In most cases, as.df = TRUE will make the most sense. (By default: as.df = TRUE)

max.hits

A single integer value specifying the maximum number of results to return or FALSE. The actual number of results will actually be higher than max.hits, as clipping only occurs on pagination page boundaries. To disable the limit, set max.hits = NULL. (By default: max.hits = 200)

use.cache

A single boolean value specifying whether to cache the result -- and reuse any existing cache entry instead of issuing a new callout. In general the use of caching for queries is discouraged, as new data is being uploaded to MGnify all the time, which might potentially be missed. However, for some purposes (such as analysis reproducibility) caching makes sense. (By default: use.cache = FALSE)

Value

A nested list or data.frame containing the results of the query.

Details

doQuery is a flexible query function, harnessing the "full" power of the JSONAPI MGnify search filters. Search results may be filtered by metadata value, associated study/sample/analyese etc. Details of the capabilities may be found [here](https://emg-docs.readthedocs.io/en/latest/api.html#customising-queries). Currently, the following filters are available (based on examination of the Python source code):

studies: accession, biome_name, lineage, centre_name

samples: accession, experiment_type, biome_name, lineage, geo_loc_name, latitude_gte, latitude_lte, longitude_gte, longitude_lte, species, instrument_model, instrument_platform, metadata_key, metadata_value_gte, metadata_value_lte, metadata_value, environment_material, environment_feature, study_accession

runs: accession, experiment_type, biome_name, lineage, species, instrument_platform, instrument_model

Unfortunately it appears that in some cases, some of these filters don't work as expected, so it is important to check the results returned match up with what's expected. Even more unfortunately if there's an error in the parameter specification, the query will run as if no filter parameters were present at all. Thus the result will appear superficially correct but will infact correspond to something completely different. This behaviour will hopefully be fixed in future incarnations of the MGnifyR or JSONAPI, but for now users should double check returned values.

It is currently not possible to combine queries of the same type in a single call (for example to search for samples between latitude). However, it is possible to run multiple queries and combine the results using set operations in R to get the desired behaviour.

Examples

mg <- MgnifyClient(cache_dir="/tmp/mgcache")

# Get a list of studies from the Agricultural Wastewater :
agwaste_studies <- doQuery(
    mg, "studies", biome_name="Agricultural wastewater"
    )

# Get all samples from a particular study
samps <- doQuery(mg, "samples", study_accession="MGYS00004521")

# Search polar samples
samps_np <- doQuery(mg, "samples", latitude_gte=66, max.hits=10)
samps_sp <- doQuery(mg, "samples", latitude_lte=-66, max.hits=10)

# Search studies that have studied drinking water
tbl <- doQuery(
    mg,
    type = "studies",
    biome_name = "root:Environmental:Aquatic:Freshwater:Drinking water",
    max.hits = 10)